+++ src/core/base/index.js
@@ -6,3 +6,4 @@ export { default as getFeatureStrAndStpMapping } from './get-feature-str-and-stp
 export { default as handleModuleSharedToGlobalMapping } from './handle-module-shared-to-global-mapping';
 export { default as runLater } from './run-later';
 export { default as uuid } from './uuid';
+export { default as buildMockEvent } from './build-mock-event';